Split macros from utils
