+++ src/ops/create_bucket.rs
@@ -12,7 +12,8 @@ use crate::headers::{
 };
 use crate::output::S3Output;
 use crate::storage::S3Storage;
-use crate::utils::{deserialize_xml_body, ResponseExt};
+use crate::utils::body::deserialize_xml_body;
+use crate::utils::ResponseExt;
 use crate::{async_trait, Method, Response};
 
 /// `CreateBucket` handler
